requirements.txt
streamlit
requests 
python-dotenv
langchain
langchain-community
openai
langchain-core
faiss-cpu
pypdf
tiktoken
langchain-openai

